Move makeCurrentMigrationRunner into current.ts
